@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.4+4
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.18.5 <3.0.0"
 
 dependencies:
   flutter:
@@ -26,40 +26,36 @@ dependencies:
   flutter_localizations:
     sdk: flutter
 
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  firebase_auth: ^0.18.0+1
-  shared_preferences: any
-  cloud_firestore: ^0.14.0+2
-  popup_menu: ^1.0.5
+
+  cupertino_icons:
+  shared_preferences: ^2.1.0
+  cloud_firestore: ^4.6.0
   flutter_dropdown: ^0.0.8
-  firebase_core: ^0.5.0
-  firebase_storage: ^4.0.0
-  image_picker: ^0.6.7+7
-  uuid: 2.2.2
-  easy_localization: ^2.3.3
-  flutter_spinkit: ^4.1.2+1
-  flutter_phone_direct_caller: ^1.0.1
-  custom_switch: ^0.0.1
-  provider: any
-  fluttertoast: any
-  #firebase_storage: ^8.0.0
-  #intl: ^0.17.0
-  carousel_slider: ^2.2.1
-  cached_network_image: ^2.1.0
-  barcode_widget:
-  pinput:
-  firebase_messaging:
-  advance_pdf_viewer:
-  flutter_cached_pdfview:
-  url_launcher:
-  mailer:
-  flutter_email_sender:
+  firebase_core: ^2.11.0
+  firebase_storage: ^11.1.2
+  image_picker: ^0.8.7+4
+  easy_localization: ^3.0.1
+  flutter_spinkit: ^5.2.0
+  flutter_phone_direct_caller: ^2.1.1
+  provider: ^6.0.5
+  fluttertoast: ^8.2.1
+  carousel_slider: ^4.2.1
+  cached_network_image: ^3.2.3
+  barcode_widget: ^2.0.3
+  pinput: ^2.2.31
+  firebase_messaging: ^14.5.0
+  firebase_auth: ^4.5.0
+  advance_pdf_viewer_fork: ^2.0.0
+  flutter_cached_pdfview: ^0.4.1
+  url_launcher: ^6.1.10
+  flutter_email_sender: ^5.2.0
+  photo_view: ^0.14.0
+  flutter_local_notifications: ^14.0.0+1
+  flutter_swiper_null_safety: ^1.0.2
   flutter_barcode_scanner:
   syncfusion_flutter_barcodes:
-  photo_view: ^0.9.2
-  flutter_local_notifications:
+
+
 
 dev_dependencies:
   flutter_test:
