@@ -1,5 +1,6 @@
-package com.sunpower.guideware
+package sunpower.sunpower
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity()
+class MainActivity: FlutterActivity() {
+}
