@@ -18,7 +18,7 @@ void main() async{
   WidgetsFlutterBinding.ensureInitialized();
   await Firebase.initializeApp();
   await LocalStorageService.instance.init();
-  notificationHelper.firebaseCloudMessagingListeners();
+  notificationHelper.firebaseCloudMessaging_Listeners();
   if(LocalStorageService.instance.languageCode == null){
     LocalStorageService.instance.languageCode = "en";
   }
@@ -50,9 +50,9 @@ class MyApp extends StatelessWidget {
                 if (AppLocalizations.of(context).locale.languageCode ==
                     "ku") {
                   child = Directionality(
-                      textDirection: TextDirection.rtl, child: child);
+                      textDirection: TextDirection.rtl, child: child!);
                 }
-                return child;
+                return child!;
               },
               localizationsDelegates: [
                 const AppLocalizationsDelegate(),
@@ -64,7 +64,7 @@ class MyApp extends StatelessWidget {
                 KurdishCupertinoLocalization.delegate
               ],
               supportedLocales: Lang.values.map((e) => Locale(e)).toList(),
-              locale: Locale(LocalStorageService.instance.languageCode),
+              locale: Locale(LocalStorageService.instance.languageCode??"en"),
               //Test()
             );
           },
